move to the given output_dtype

+++ graphbolt/src/csc_to_coo.cc
@@ -28,7 +28,9 @@ torch::Tensor CSCToCOO(
     original_column_node_ids =
         torch::arange(indptr.size(0) - 1, indptr.options().dtype(output_dtype));
   }
-  return original_column_node_ids.value().repeat_interleave(indptr.diff());
+  return original_column_node_ids.value()
+      .repeat_interleave(indptr.diff())
+      .to(output_dtype);
 }
 
 }  // namespace ops
